* dev: (84 commits)
  [book]
  [build] preparing 0.12.01a
  [map] disabled DebugLog in Update()
  [db] updated PlaySession (now 2 PS per LB)
  [sfx] new Antura Sfx.Noize
  [Credits] get loaded from Credits.txt file
  [MapScene] Add pin stage 6 + fixed some bugs
  [ThrowBalls] Fix word not showing on WordFlexibleContainer in Words variation.
  [ThrowBalls] Antura opening sequence.
  [icon] new App Icon
  [icon] new App Icon
  [clean] deleted old boo kscene
  [map camera] use shorter sfx
  [fixed MapScene] First Contact UI
  [fixed MapScene] First Contact
  [Balloons] no star flowers
  [clean]
  [GameSelector] Back button instead of Pause menu
  [sfx] trimmed blip
  [AnturaSpace][change]Added background music, reapplying lost changes
  ...

+++ Assets/_app/_scripts/TeacherAI/AIEngines/LogAI.cs
@@ -123,7 +123,7 @@ namespace EA4S.Teacher
 
         public void LogMiniGameScore(MiniGameCode miniGameCode, float score)
         {
-            UpdateScoreDataWithMaximum(DbTables.MiniGames, ((int)miniGameCode).ToString(), score);
+            UpdateScoreDataWithMaximum(DbTables.MiniGames, (miniGameCode).ToString(), score);
         }
 
         public void LogPlaySessionScore(string playSessionId, float score)
